Added proper stack traces for non serialized errors from libraries
