+++ apps/ta-portal/ui/src/services/db-apis.js
@@ -20,6 +20,8 @@ async function handleApiResponse(response) {
       .catch(() => ({ message: "Unknown error" }));
     console.log("API error body:", errorBody);
 
+    // Prefer a concise 'name' from the backend when available
+    const conciseName = errorBody.name || null;
     const errorMessage =
       errorBody.error ||
       errorBody.message ||
@@ -28,6 +30,8 @@ async function handleApiResponse(response) {
     sessionStorage.setItem(
       "errorDetails",
       JSON.stringify({
+        // Use name if available (short label), else fall back to the concise message
+        name: conciseName,
         error: errorMessage,
         statusCode: response.status,
         url: response.url,
@@ -39,9 +43,10 @@ async function handleApiResponse(response) {
             : new Error().stack, // fallback frontend trace
       })
     );
-    
+
     window.location.href = "/Error";
-    throw new Error(errorMessage);
+    // Throw the concise name (if present) or the message to keep console errors readable
+    throw new Error(conciseName || errorMessage);
   }
   return response.json();
 }
